backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
boto3==1.35.93
requests-oauthlib==2.0.0
cryptography==44.0.0
python-dotenv==1.0.1
pymongo==4.8.0
pydantic==2.10.6
email-validator==2.2.0
pyjwt==2.10.1
bcrypt==4.2.1
passlib==1.7.4
tzdata==2025.2
motor==3.5.1
pytest==8.3.4
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.14.1
python-jose==3.3.0
requests==2.32.3
python-multipart==0.0.20
jq==1.8.0
typer==0.15.1
emergentintegrations==0.1.0
litellm==1.56.8
beautifulsoup4==4.14.0
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
lxml==5.3.0
soupsieve==2.6
httpx==0.27.2
aiohttp==3.11.11
openai==1.99.9
fastuuid==0.14.0
PyYAML==6.0.2
jinja2==3.1.5
anthropic==0.42.0
google-generativeai==0.8.4
tiktoken==0.8.0
tokenizers==0.21.0
psutil==6.1.1
discord.py==2.4.0
redis==7.1.0
celery==5.6.0


razorpay==1.4.2
supabase==2.10.0
